Move IdentityTransformStream/FixedLengthStream to its own files
Solely for the purpose of cleaning up internal.h/c++ and reducing
the size of that file in preparation for further cleanups.

Zero functional changes are made beyond moving code around.
The IdentityTransformStreamImpl class is moved into the
compilation file identity-transform-stream.c++ and its
functions are defined inline in the class, but otherwise
the implementation is unchanged.
