+++ src/workerd/api/streams.h
@@ -9,6 +9,7 @@
 
 #include "streams/compression.h"
 #include "streams/encoding.h"
+#include "streams/identity-transform-stream.h"
 #include "streams/readable.h"
 #include "streams/standard.h"
 #include "streams/transform.h"
