+++ src/workerd/api/streams/internal.h
@@ -370,75 +370,4 @@ class WritableStreamInternalController: public WritableStreamController {
   // events.
   std::list<WriteEvent> queue;
 };
-
-// An implementation of ReadableStreamSource and WritableStreamSink which communicates read and
-// write requests via a OneOf.
-//
-// This class is also used as the implementation of FixedLengthStream, in which case `limit` is
-// non-nullptr.
-class IdentityTransformStreamImpl: public kj::Refcounted,
-                                   public ReadableStreamSource,
-                                   public WritableStreamSink {
-  // TODO(soon): Reimplement this in terms of kj::OneWayPipe, so we can optimize pumpTo().
-
- public:
-  explicit IdentityTransformStreamImpl(kj::Maybe<uint64_t> limit = kj::none): limit(limit) {}
-
-  ~IdentityTransformStreamImpl() noexcept(false) {
-    // Due to the different natures of JS and C++ disposal, there is no point in enforcing the limit
-    // for a FixedLengthStream here.
-    //
-    // 1. Creating but not using a `new FixedLengthStream(n)` should not be an error, and ought not
-    //    to logspam us.
-    // 2. Chances are high that by the time this object gets destroyed, it's too late to tell the
-    //    user about the failure.
-  }
-
-  // ReadableStreamSource implementation -------------------------------------------------
-
-  kj::Promise<size_t> tryRead(void* buffer, size_t minBytes, size_t maxBytes) override;
-
-  kj::Promise<size_t> tryReadInternal(void* buffer, size_t maxBytes);
-
-  kj::Promise<DeferredProxy<void>> pumpTo(WritableStreamSink& output, bool end) override;
-
-  kj::Maybe<uint64_t> tryGetLength(StreamEncoding encoding) override;
-
-  void cancel(kj::Exception reason) override;
-
-  // WritableStreamSink implementation ---------------------------------------------------
-
-  kj::Promise<void> write(kj::ArrayPtr<const byte> buffer) override;
-
-  kj::Promise<void> write(kj::ArrayPtr<const kj::ArrayPtr<const kj::byte>> pieces) override;
-
-  kj::Promise<void> end() override;
-
-  void abort(kj::Exception reason) override;
-
- private:
-  kj::Promise<size_t> readHelper(kj::ArrayPtr<kj::byte> bytes);
-
-  kj::Promise<void> writeHelper(kj::ArrayPtr<const kj::byte> bytes);
-
-  kj::Maybe<uint64_t> limit;
-
-  struct ReadRequest {
-    kj::ArrayPtr<kj::byte> bytes;
-    // WARNING: `bytes` may be invalid if fulfiller->isWaiting() returns false! (This indicates the
-    //   read was canceled.)
-
-    kj::Own<kj::PromiseFulfiller<size_t>> fulfiller;
-  };
-
-  struct WriteRequest {
-    kj::ArrayPtr<const kj::byte> bytes;
-    kj::Own<kj::PromiseFulfiller<void>> fulfiller;
-  };
-
-  struct Idle {};
-
-  kj::OneOf<Idle, ReadRequest, WriteRequest, kj::Exception, StreamStates::Closed> state = Idle();
-};
-
 }  // namespace workerd::api
